requirements.txt
opencv-python==4.8.1.78
openai==1.12.0
python-dotenv==1.0.0
python-telegram-bot==20.7
numpy==1.24.3
Pillow==10.0.1
rich==13.7.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
gradio>=5.31.0
pre-commit==4.2.0
